+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Initial clone records script
 * Initial set resource IDs script
+* Initial update collections script
 
 ### Changed